Bring SYCL allocators under MemManage.
