+++ src/Platforms/DeviceAllocator.hpp
@@ -21,9 +21,9 @@
 #include <config.h>
 
 #if (defined(ENABLE_CUDA) || defined(ENABLE_SYCL)) && !defined(ENABLE_OFFLOAD)
+#include "MemManageAlias.hpp"
 
 #if defined(ENABLE_CUDA)
-#include "MemManageAlias.hpp"
 namespace qmcplusplus
 {
 template<typename T>
@@ -31,7 +31,6 @@ using DeviceAllocator = CUDAAllocator<T>;
 }
 
 #elif defined(ENABLE_SYCL)
-#include "SYCL/SYCLallocator.hpp"
 namespace qmcplusplus
 {
 template<typename T>
