+++ src/Platforms/SYCL/CMakeLists.txt
@@ -9,7 +9,7 @@
 #// File created by: Ye Luo, yeluo@anl.gov, Argonne National Laboratory
 #//////////////////////////////////////////////////////////////////////////////////////
 
-set(SYCL_RT_SRCS SYCLDeviceManager.cpp SYCLallocator.cpp SYCLruntime.cpp)
+set(SYCL_RT_SRCS SYCLDeviceManager.cpp MemManageSYCL.cpp SYCLruntime.cpp)
 set(SYCL_LA_SRCS syclBLAS.cpp)
 
 add_library(platform_sycl_runtime ${SYCL_RT_SRCS})
@@ -17,6 +17,8 @@ target_link_libraries(
   platform_sycl_runtime
   PUBLIC SYCL::host
   PRIVATE platform_host_runtime)
+
+target_include_directories(platform_sycl_runtime PRIVATE "${CMAKE_CURRENT_SOURCE_DIR}/..")
 
 add_library(platform_sycl_LA ${SYCL_LA_SRCS})
 target_link_libraries(
