+++ src/Platforms/SYCL/MemManageSYCL.cpp
@@ -2,7 +2,7 @@
 // This file is distributed under the University of Illinois/NCSA Open Source License.
 // See LICENSE file in top directory for details.
 //
-// Copyright (c) 2022 QMCPACK developers.
+// Copyright (c) 2025 QMCPACK developers.
 //
 // File developed by: Ye Luo, yeluo@anl.gov, Argonne National Laboratory
 //
@@ -10,10 +10,15 @@
 //////////////////////////////////////////////////////////////////////////////////////
 
 
+#include "MemManageSYCL.hpp"
 #include <cstddef>
 #include <atomic>
 
 namespace qmcplusplus
 {
 std::atomic<size_t> SYCLallocator_device_mem_allocated(0);
-}
+namespace compute
+{
+template class MemManage<PlatformKind::SYCL>;
+} // namespace compute
+} // namespace qmcplusplus
