+++ src/QMCWaveFunctions/Fermion/syclSolverInverter.hpp
@@ -14,7 +14,7 @@
 
 #include "OhmmsPETE/OhmmsVector.h"
 #include "OhmmsPETE/OhmmsMatrix.h"
-#include "SYCL/SYCLallocator.hpp"
+#include "SYCL/MemManageSYCL.hpp"
 #include "SYCL/syclBLAS.hpp"
 #include "SYCL/syclSolver.hpp"
 #include "QMCWaveFunctions/detail/SYCL/sycl_determinant_helper.hpp"
